requirements.txt
dbt-bigquery==1.10.1

google-cloud-secret-manager
google-cloud-bigquery
google-cloud-functions
requests
pyyaml
flask
flask-cors==4.0.1
python-dotenv
gunicorn
